With_Python/requirements.txt
moviepy==1.0.3 #https://pypi.org/project/moviepy/#history
pymediainfo==6.1.0


### Speech

### Photo
pillow==11.0.0